Add custom snyk init script to handle buildScript dependencies (#7)
Description
===========

Snyk can not report on `buildScript` dependencies like plugins and custom
build time libraries. I created a custom script which overrides the default script
from snyk which is included in the snyk cli. The only way of replacing it is to
force override the script. I don't want to handle this in the main plugin as
this is a pretty dirty workaround which might break in the future.

Changes
=======

* ![ADD] convention for a custom `init-scritp` to report buildScript dependencies